+++ front/src/components/atoms/Button.svelte
@@ -1,6 +1,12 @@
+<script lang=ts>
+  export let primary = false;
+
+  $: mode = primary ? 'button--primary' : 'button--secondary';
+</script>
+
 <button
   on:click
-  class="button text-preset-4"
+  class="{['button text-preset-4', mode].join(' ')}"
 >            
     <slot/>
 </button>
