--- front/src/stories/Header.stories.ts
@@ -1,26 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/svelte';
-import Header from './Header.svelte';
-
-const meta = {
-  title: 'Example/Header',
-  component: Header,
-  // This component will have an automatically generated Autodocs entry: https://storybook.js.org/docs/react/writing-docs/autodocs
-  tags: ['autodocs'],
-  parameters: {
-    // More on how to position stories at: https://storybook.js.org/docs/svelte/configure/story-layout
-    layout: 'fullscreen',
-  },
-} satisfies Meta<Header>;
-
-export default meta;
-type Story = StoryObj<typeof meta>;
-
-export const LoggedIn: Story = {
-  args: {
-    user: {
-      name: 'Jane Doe',
-    },
-  },
-};
-
-export const LoggedOut: Story = {};
