+++ front/src/stories/molecules/Mol.svelte
@@ -1,0 +1,6 @@
+<script>
+  import Button from '../../components/atoms/Button.svelte';
+  import '../../assets/css/global.css';  
+</script>
+
+<Button>Button</Button>
