+++ front/src/stories/organisms/Org.stories.ts
@@ -1,0 +1,27 @@
+import type { Meta, StoryObj } from '@storybook/svelte';
+
+import Button from './Org.svelte';
+
+// More on how to set up stories at: https://storybook.js.org/docs/svelte/writing-stories/introduction
+const meta = {
+  title: 'Organismes/Organisme name',
+  component: Button,
+  tags: ['autodocs'],
+} satisfies Meta<Button>;
+
+export default meta;
+type Story = StoryObj<typeof meta>;
+
+// More on writing stories with args: https://storybook.js.org/docs/svelte/writing-stories/args
+export const Primary: Story = {
+  args: {
+    primary: true,
+    label: 'Button',
+  },
+};
+
+export const Secondary: Story = {
+  args: {
+    label: 'Button',
+  },
+};
